Separated RecordKind into separate file
